Move Cluster page components into their own folder
