+++ assets/src/components/Console.tsx
@@ -18,13 +18,14 @@ import Webhooks from './Webhooks'
 import { EnsureLogin } from './Login'
 import Users from './Users'
 import { InstallationsProvider } from './Installations'
-import { Pod } from './apps/app/components/kubernetes/pods/Pod'
-import Pods from './apps/app/components/kubernetes/pods/Pods'
-import Node from './apps/app/components/kubernetes/nodes/Node'
-import Nodes from './apps/app/components/kubernetes/nodes/Nodes'
-import NodeInfo from './apps/app/components/kubernetes/nodes/NodeInfo'
-import NodeEvents from './apps/app/components/kubernetes/nodes/NodeEvents'
-import NodeRaw from './apps/app/components/kubernetes/nodes/NodeRaw'
+import Cluster from './cluster/Cluster'
+import { Pod } from './cluster/pods/Pod'
+import Pods from './cluster/pods/Pods'
+import Node from './cluster/nodes/Node'
+import Nodes from './cluster/nodes/Nodes'
+import NodeInfo from './cluster/nodes/NodeInfo'
+import NodeEvents from './cluster/nodes/NodeEvents'
+import NodeRaw from './cluster/nodes/NodeRaw'
 import Directory from './users/Directory'
 import EditUser from './users/EditUser'
 import { Audits } from './audits/Audits'
@@ -54,7 +55,6 @@ import Component from './apps/app/components/component/Component'
 import ComponentInfo from './apps/app/components/component/ComponentInfo'
 import ComponentEvents from './apps/app/components/component/ComponentEvents'
 import ComponentRaw from './apps/app/components/component/ComponentRaw'
-import Cluster from './apps/app/components/kubernetes/Cluster'
 
 export const TOOLBAR_HEIGHT = '55px'
 export const SIDEBAR_WIDTH = '200px'
