+++ assets/src/components/cluster/StatefulSet.jsx
