+++ assets/src/components/cluster/nodes/ClusterGauges.tsx
