+++ assets/src/components/cluster/nodes/NodeRaw.tsx
