+++ assets/src/components/cluster/nodes/NodeSideNav.tsx
