+++ assets/src/components/cluster/nodes/NodeSidecar.tsx
